+++ src/AdoNet/Shared/Storage/DbStoredQueries.cs
@@ -150,7 +150,7 @@ namespace Orleans.Tests.SqlUtils
 
                         //Use the GetInt32 method instead of the generic GetValue<TValue> version to retrieve the value from the data record
                         //GetValue<int> causes an InvalidCastException with oracle data provider. See https://github.com/dotnet/orleans/issues/3561
-                        Period = TimeSpan.FromMilliseconds(record.GetInt32(nameof(Columns.Period))),
+                        Period = TimeSpan.FromMilliseconds(record.GetInt64(nameof(Columns.Period))),
                         ETag = GetVersion(record).ToString()
                     };
                 }
@@ -398,7 +398,7 @@ namespace Orleans.Tests.SqlUtils
 
             internal TimeSpan Period
             {
-                set { Add(nameof(Period), (int)value.TotalMilliseconds); }
+                set { Add(nameof(Period), (long)value.TotalMilliseconds); }
             }
 
             internal SiloStatus Status
